feat: split python and rattler-build protocol into seperate files
