+++ crates/pixi-build/src/bin/pixi-build-python/main.rs
@@ -1,12 +1,13 @@
 mod build_script;
 mod config;
+mod protocol;
 mod python;
 
-use python::PythonBuildBackend;
+use protocol::PythonBuildBackendFactory;
 
 #[tokio::main]
 pub async fn main() {
-    if let Err(err) = pixi_build_backend::cli::main(PythonBuildBackend::factory).await {
+    if let Err(err) = pixi_build_backend::cli::main(PythonBuildBackendFactory::new).await {
         eprintln!("{err:?}");
         std::process::exit(1);
     }
