+++ crates/pixi-build/src/bin/pixi-build-rattler-build/main.rs
@@ -1,10 +1,11 @@
+mod protocol;
 mod rattler_build;
 
-use rattler_build::RattlerBuildBackend;
+use protocol::RattlerBuildBackendFactory;
 
 #[tokio::main]
 pub async fn main() {
-    if let Err(err) = pixi_build_backend::cli::main(RattlerBuildBackend::factory).await {
+    if let Err(err) = pixi_build_backend::cli::main(RattlerBuildBackendFactory::new).await {
         eprintln!("{err:?}");
         std::process::exit(1);
     }
